#192 [mod]: TermsNavigation 삭제 -> SplashNavigation으로 이동
